% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.r
\name{trim_edges}
\alias{trim_edges}
\title{Trim the low-weight edges from a graph}
\usage{
trim_edges(x, k, method = "top_k")
}
\arguments{
\item{x}{a subgroup object.}

\item{k}{parameter for the trim method.}

\item{method}{the method for trimming edges. Current options are the
default, "top_k" which retains the edges with the k highest values or
"cutoff" which is value a weight must be above to remain in the graph.}
}
\description{
Trim the low-weight edges from a graph
}
