% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.r
\name{trim_edges}
\alias{trim_edges}
\title{Trim the low-weight edges from a graph}
\usage{
trim_edges(x, k, method = "top_k")
}
\arguments{
\item{x}{a subgroup object.}

\item{k}{parameter for the trim method. Currently, it can be either
"top_k", which keeps the edges with the k largest weights or 
"cutoff" where edges whose weights are less than the cutoff are removed.}

\item{method}{the method for trimming edges. Current options are the
default, "top_k" which retains the edges with the k highest values or
"cutoff" which is value a weight must be above to remain in the graph.}
}
\description{
Remove edges in a graph whose weights are small. This function
is meant to visualize full-connected (or almost fully connected graphs) 
where many of the edges may be small.
}
